Update README.md
it felt like something i should be really good at, and i had high ambitions from the start, i envisioned new ways of experiencing news reading, i had a data visualization project ahead of it's time that i was working on i called nexii, and i stayed up late nights in high school trying to make it come about

there was something precious that i felt i could manifest in the world. i was in love with the potential and possibilities of the future. i am again, but it has been quite a journey....